Move enums and fix translations
