+++ app/Filament/Clusters/Databases/Pages/DatabaseType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models\Enums;
+namespace App\Filament\Clusters\Databases\Pages;
 
 enum DatabaseType: string
 {
