+++ app/Services/Database/DatabaseFactory.php
@@ -2,7 +2,7 @@
 
 namespace App\Services\Database;
 
-use App\Models\Enums\DatabaseType;
+use App\Filament\Clusters\Databases\Pages\DatabaseType;
 use App\Services\Database\Vendors\MySQL;
 use App\Services\Database\Vendors\SQLite;
