+++ tests/Unit/Config/AppTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Models\Enums\DatabaseType;
+use App\Filament\Clusters\Databases\Pages\DatabaseType;
 
 it('defines correct environment for test', function (): void {
     expect(env('APP_ENV'))->toBe('testing');
